pyproject.toml
[project]
name = "mojofs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "blake3>=1.0.5",
    "brotli>=1.1.0",
    "deflate>=0.8.1",
    "highwayhash-cffi>=0.1.6",
    "lz4>=4.4.4",
    "msgpack-python>=0.5.6",
    "python-snappy>=0.7.3",
    "rmp>=0.0.0.post20250826141809",
    "siphashc>=2.5",
    "zstd>=1.5.7.2",
]

[tool.setuptools]
include-package-data = false

[tool.setuptools.packages.find]
include = ['mojofs']
exclude = []

[tool.setuptools.package-data]


[[tool.uv.index]]
url = "https://repo.huaweicloud.com/repository/pypi/simple"
default = true
[tool.pixi.workspace]
channels = ["conda-forge"]
platforms = ["linux-64"]

[tool.pixi.pypi-dependencies]
mojofs = { path = ".", editable = true }

[tool.pixi.tasks]
